test: common.test testclient refactored further with common methods from CasedbAppTestClient
