+++ test/casedb/integration/build_db/delete.py
@@ -1,4 +1,4 @@
-from test.casedb.casedb_app_test_client import CasedbAppTestClient as Env
+from test.casedb.casedb_app_test_client import CasedbTestClient as Env
 from test.casedb.integration.build_db.base import (
     APP_ADMIN_OR_ABOVE_USERS,
     BELOW_APP_ADMIN_DATA_USERS,
