+++ test/casedb/integration/build_db/read.py
@@ -1,4 +1,4 @@
-from test.casedb.casedb_app_test_client import CasedbAppTestClient as Env
+from test.casedb.casedb_app_test_client import CasedbTestClient as Env
 from test.casedb.integration.build_db.base import (
     ALL_USERS,
     DATA_USERS,
