+++ test/casedb/integration/case_access/test_case_access.py
@@ -1,7 +1,7 @@
 import logging
 import pickle
 from pathlib import Path
-from test.casedb.casedb_app_test_client import CasedbAppTestClient as Env
+from test.casedb.casedb_app_test_client import CasedbTestClient as Env
 from test.casedb.integration.case_access.base import (
     REPOSITORY_TYPE,
     SKIP_ENDPOINTS,
